// packages/destination-actions/src/destinations/klime/identify/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * Unique identifier for the event
   */
  messageId: string
  /**
   * Unique identifier for the user
   */
  userId: string
  /**
   * Unique identifier for the group/organization
   */
  groupId?: string
  /**
   * User attributes (email, name, plan, etc.)
   */
  traits?: {
    [k: string]: unknown
  }
  /**
   * When the event occurred (ISO 8601)
   */
  timestamp?: string | number
  /**
   * Contextual information about the event
   */
  context?: {
    [k: string]: unknown
  }
  /**
   * When enabled, events are sent in batches.
   */
  enable_batching: boolean
  /**
   * Maximum number of events per batch request.
   */
  batch_size?: number
}